% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfclean.R
\name{dfclean}
\alias{dfclean}
\title{Clean and organize data to prepare for calculating indices of inequality}
\usage{
dfclean(ineqvar, df, order, offsetvar)
}
\arguments{
\item{ineqvar}{variable name, in quotations, of the variable within df measuring inequality}

\item{df}{the relevant input dataframe}

\item{order}{either "l2h" for low to high or "h2l" for high to low based on how the ineqvar needs to be arranged}

\item{offsetvar}{the count variable from df, in quotations, used for creating the frequencies and the ridit}
}
\value{
dataframe with the new variables and data arranged in the specified order
}
\description{
This function aims to rearrange the data in a dataframe by the variable measuring inequality (ineqvar) in a specified order.
Additionally, it creates new variables within the dataframe for frequency, cumulative frequency, and ridit.
}
