% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfexample.R
\docType{data}
\name{dfexample}
\alias{dfexample}
\title{Example data from Colombia}
\format{
A data frame with 5 variables: \code{municipios}, \code{MN},
\code{NV}, \code{TMN}, \code{IDSx}
\describe{
\item{municipios}{The name of the municipality}
\item{MN}{The number of stillbirths}
\item{NV}{The number of live births}
\item{TMN}{The rate of stillbirths per 1000 births}
\item{IDSx}{Sustainable Development Index, a measure of social inequality}
}
}
\usage{
dfexample
}
\description{
Source: Pan American Health Organization
}
\keyword{datasets}
