% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riditreg.R
\name{riditreg}
\alias{riditreg}
\title{Create models and calculate indices of inequality}
\usage{
riditreg(df, yvar, method, offsetvar, countvar, conf_level = 0.95)
}
\arguments{
\item{df}{the input dataframe organized correctly by previously using the dfclean() function}

\item{yvar}{the rate variable name, in quotations, used as the response variable in some of the model options}

\item{method}{the regression method to be used with options "lm", "poisson", "nb", and "logistic"}

\item{offsetvar}{the count variable, in quotations, of the total number of outcomes in each group used as weights/offsets in the models}

\item{countvar}{the count variable, in quotations, of the number of occurrences in each group used as counts in the logistic regression}

\item{conf_level}{confidence level for the confidence intervals}
}
\value{
A list consisting of the following:
\item{regression}{The summary output of the selected regression model.}
\item{SII_01}{The Slope Index of Inequality comparing ridit=0 and ridit=1.}
\item{RII_01}{The Relative Index of Inequality comparing ridit=0 and ridit=1.}
\item{SII_minmax}{The Slope Index of Inequality comparing the min ridit value from df and the max ridit value from df.}
\item{RII_minmax}{The Relative Index of Inequality comparing the min ridit value from df and the max ridit value from df.}
}
\description{
This function aims to take the cleaned/organized data from the dfclean() function and returns relevant regression models.
It also returns estimates of the Slope Index of Inequality and Relative Index of Inequality for the comparison of both
the least and greatest ridit values in the dataframe as well as the theoretical min and max of 0 and 1.
}
